+++ lib/src/function_generator_helper.dart
@@ -1,0 +1,55 @@
+import 'package:analyzer/dart/element/element.dart';
+import 'package:recase/recase.dart';
+import 'package:riverpod_mutations_generator/src/class_generator_helper.dart';
+import 'package:riverpod_mutations_generator/src/util.dart';
+
+class FunctionGeneratorHelper {
+  static String mutationsForFunction(FunctionElement function) {
+    // return '/* Hi $function */';
+    // return providerForFunction(function); // null error
+    // return mutationClassesFromFunction(function);
+    return '''
+${providerForFunction(function)}
+${mutationClassesFromFunction(function)}
+''';
+  }
+
+  static String mutationClassesFromFunction(FunctionElement function) {
+    return ClassGeneratorHelper.mutationClasses(function);
+  }
+
+  static String providerForFunction(FunctionElement function) {
+    final String accessName = function.name;
+    // switch (function) {
+    //   // TODO: test if this works right
+    //   FunctionElement(isStatic: true) =>
+    //     '${function.enclosingElement.name!}.${function.name}',
+    //   FunctionElement(isStatic: false) => function.name,
+    // };
+    if (!function.parameters.any(mutationKeyTypeChecker.hasAnnotationOf)) {
+      return '''
+final ${function.name}Provider = Provider.autoDispose((ref) {
+  return ${function.name.pascalCase}Mutation(
+    (newState) => ref.state = newState,
+    ${accessName},
+  );
+});
+''';
+    }
+    final familyParameters = Util.parameterListToString(
+      function.parameters.where(mutationKeyTypeChecker.hasAnnotationOf),
+      removeDefaults: true,
+      removeRequired: true,
+    );
+
+    return '''
+typedef ${function.name.pascalCase}FamilyParameters = (${familyParameters});
+final ${function.name}Provider = Provider.autoDispose.family((ref, ${function.name.pascalCase}FamilyParameters params) {
+  return ${function.name.pascalCase}Mutation(
+    (newState) => ref.state = newState..params = params,
+    ${accessName},
+  )..params = params;
+});
+''';
+  }
+}